backend/requirements.txt
Flask
opencv-python
tensorflow==2.19.0
Pillow
numpy
flask-cors